+++ src/app/app.component.html
@@ -1,5 +1,3 @@
-<!-- <router-outlet></router-outlet> -->
-
 <div class="wrapper">
 	<div class="box header">
 			<app-header></app-header>
