--- src/app/app.routing.1.ts
@@ -1,19 +1,0 @@
-import { ModuleWithProviders } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { HomeComponent } from './components/home/home.component';
-import { ProfileComponent } from './components/profile/profile.component';
-
-const appRoutes: Routes = [
-    {
-        path: '',
-        component: HomeComponent
-    }, {
-        path: 'profile',
-        component: ProfileComponent
-    }
-];
-
-export const appRoutingProviders: any[] = [];
-
-export const routing: ModuleWithProviders = RouterModule.forRoot(appRoutes);
