--- src/app/components/home1/home.component.html
@@ -1,1 +1,0 @@
-<h3>Welcome Home</h3>
