--- src/app/components/home1/home.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-    // moduleId: module.id,
-    selector: 'app-home',
-    templateUrl: './home.component.html',
-    styleUrls: ['./home.component.css']
-})
-
-export class HomeComponent implements OnInit {
-
-    constructor() { }
-
-    ngOnInit() {
-    }
-
-}
