+++ src/app/home/home.component.html
@@ -1,0 +1,3 @@
+<p>
+  home works!
+</p>
